require.js : initial add and refractor to modules.
I converted the application code and basic plugin features to require.js.

The basic code was easy to execute, but dynamically loading jQuery plugins
was extremely error prone.  I need to make a general purpose loader that
can be chained in a functional style, since this is jQuery and requires
preffered technique (no blocking, only delegation).

Signed-off-by: James Arlow <james.arlow@gmail.com>